+++ nicegui/nicegui.py
@@ -1,6 +1,5 @@
 import asyncio
 import mimetypes
-import time
 import urllib.parse
 from pathlib import Path
 from typing import Dict
@@ -11,19 +10,19 @@ from fastapi.responses import FileResponse, Response
 from fastapi.staticfiles import StaticFiles
 from fastapi_socketio import SocketManager
 
-from . import (background_tasks, binding, favicon, globals, json, outbox,  # pylint: disable=redefined-builtin
-               run_executor, welcome)
+from . import background_tasks, globals, json  # pylint: disable=redefined-builtin
 from .app import App
 from .client import Client
 from .dependencies import js_components, libraries
 from .error import error_content
-from .helpers import is_file, safe_invoke
+from .helpers import safe_invoke
 from .json import NiceGUIJSONResponse
+from .lifespan import lifespan
 from .middlewares import RedirectWithPrefixMiddleware
 from .page import page
 from .version import __version__
 
-globals.app = app = App(default_response_class=NiceGUIJSONResponse)
+globals.app = app = App(default_response_class=NiceGUIJSONResponse, lifespan=lifespan)
 # NOTE we use custom json module which wraps orjson
 socket_manager = SocketManager(app=app, mount_location='/_nicegui_ws/', json=json)
 globals.sio = sio = socket_manager._sio  # pylint: disable=protected-access
@@ -67,58 +66,6 @@ def _get_component(key: str) -> FileResponse:
         headers = {'Cache-Control': 'public, max-age=3600'}
         return FileResponse(js_components[key].path, media_type='text/javascript', headers=headers)
     raise HTTPException(status_code=404, detail=f'component "{key}" not found')
-
-
-@app.on_event('startup')
-def handle_startup(with_welcome_message: bool = True) -> None:
-    """Handle the startup event."""
-    # NOTE ping interval and timeout need to be lower than the reconnect timeout, but can't be too low
-    globals.sio.eio.ping_interval = max(globals.reconnect_timeout * 0.8, 4)
-    globals.sio.eio.ping_timeout = max(globals.reconnect_timeout * 0.4, 2)
-    if not globals.ui_run_has_been_called:
-        raise RuntimeError('\n\n'
-                           'You must call ui.run() to start the server.\n'
-                           'If ui.run() is behind a main guard\n'
-                           '   if __name__ == "__main__":\n'
-                           'remove the guard or replace it with\n'
-                           '   if __name__ in {"__main__", "__mp_main__"}:\n'
-                           'to allow for multiprocessing.')
-    if globals.favicon:
-        if is_file(globals.favicon):
-            globals.app.add_route('/favicon.ico', lambda _: FileResponse(globals.favicon))  # type: ignore
-        else:
-            globals.app.add_route('/favicon.ico', lambda _: favicon.get_favicon_response())
-    else:
-        globals.app.add_route('/favicon.ico', lambda _: FileResponse(Path(__file__).parent / 'static' / 'favicon.ico'))
-    globals.state = globals.State.STARTING
-    globals.loop = asyncio.get_running_loop()
-    with globals.index_client:
-        for t in globals.startup_handlers:
-            safe_invoke(t)
-    background_tasks.create(binding.refresh_loop(), name='refresh bindings')
-    background_tasks.create(outbox.loop(), name='send outbox')
-    background_tasks.create(prune_clients(), name='prune clients')
-    background_tasks.create(prune_slot_stacks(), name='prune slot stacks')
-    globals.state = globals.State.STARTED
-    if with_welcome_message:
-        background_tasks.create(welcome.print_message())
-    if globals.air:
-        background_tasks.create(globals.air.connect())
-
-
-@app.on_event('shutdown')
-async def handle_shutdown() -> None:
-    """Handle the shutdown event."""
-    if app.native.main_window:
-        app.native.main_window.signal_server_shutdown()
-    globals.state = globals.State.STOPPING
-    with globals.index_client:
-        for t in globals.shutdown_handlers:
-            safe_invoke(t)
-    run_executor.tear_down()
-    globals.state = globals.State.STOPPED
-    if globals.air:
-        await globals.air.disconnect()
 
 
 @app.exception_handler(404)
@@ -211,52 +158,3 @@ def _on_javascript_response(_: str, msg: Dict) -> None:
 def handle_javascript_response(client: Client, msg: Dict) -> None:
     """Forward a JavaScript response to the corresponding element."""
     client.waiting_javascript_commands[msg['request_id']] = msg['result']
-
-
-async def prune_clients() -> None:
-    """Prune stale clients in an endless loop."""
-    while True:
-        try:
-            stale_clients = [
-                client
-                for client in globals.clients.values()
-                if not client.shared and not client.has_socket_connection and client.created < time.time() - 60.0
-            ]
-            for client in stale_clients:
-                _delete_client(client)
-        except Exception:
-            # NOTE: make sure the loop doesn't crash
-            globals.log.exception('Error while pruning clients')
-        await asyncio.sleep(10)
-
-
-async def prune_slot_stacks() -> None:
-    """Prune stale slot stacks in an endless loop."""
-    while True:
-        try:
-            running = [
-                id(task)
-                for task in asyncio.tasks.all_tasks()
-                if not task.done() and not task.cancelled()
-            ]
-            stale = [
-                id_
-                for id_ in globals.slot_stacks
-                if id_ not in running
-            ]
-            for id_ in stale:
-                del globals.slot_stacks[id_]
-        except Exception:
-            # NOTE: make sure the loop doesn't crash
-            globals.log.exception('Error while pruning slot stacks')
-        await asyncio.sleep(10)
-
-
-def _delete_client(client: Client) -> None:
-    """Delete a client and all its elements.
-
-    If the global clients dictionary does not contain the client, its elements are still removed and a KeyError is raised.
-    Normally this should never happen, but has been observed (see #1826).
-    """
-    client.remove_all_elements()
-    del globals.clients[client.id]
