+++ pyproject.toml
@@ -14,7 +14,7 @@ typing-extensions = ">=3.10.0"
 markdown2 = "^2.4.7"
 Pygments = ">=2.15.1,<3.0.0"
 uvicorn = {extras = ["standard"], version = "^0.22.0"}
-fastapi = ">=0.92,<1.0.0"
+fastapi = ">=0.93,<1.0.0"
 fastapi-socketio = "^0.0.10"
 python-socketio = ">=5.10.0" # https://github.com/zauberzeug/nicegui/issues/1809
 vbuild = ">=0.8.2"
